Move EnsureBot into bot.go
